`LocalFeedImageDataLoader` does not message store upon creation
